requirements.txt
alembic==1.8.1
atomicwrites==1.4.1
attrs==21.4.0
click==8.1.3
colorama==0.4.5
dataclasses-json==0.5.7
flake8==4.0.1
Flask==2.1.3
greenlet==1.1.2
iniconfig==1.1.1
Inject==4.3.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-enum==1.5.1
mccabe==0.6.1
mypy==0.971
mypy-extensions==0.4.3
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-dotenv==0.5.2
pytest-mock==3.8.2
python-dotenv==0.20.0
SQLAlchemy==1.4.39
tomli==2.0.1
typing-inspect==0.7.1
typing_extensions==4.3.0
Werkzeug==2.2.0
